refactor: excise utils/, reject BodyLen overflow at quicproto boundary
The utils/ package's five integer-conversion helpers silently clamped
overflow to a sentinel, turning data-corruption bugs into success-shaped
returns (a shard index of -1 silently overwrote shard 0). Replace all 17
call sites with bounded casts:

- ShardIndex int -> uint32 on quicserver request structs (compile-driven
  catch for negative shard indices arriving via JSON).
- DBNode.ID int -> uint64 in s3.Config (matches s3db.DBNodeConfig.ID;
  TOML decoder rejects negatives at config-load).
- NodeToUint32 now returns a real overflow error instead of silent clamp.
- BodyLen uint64 -> int64 conversions trust quicproto.ReadHeader, which
  now rejects BodyLen > MaxInt64 with ErrBodyLenOverflow.

Per-cast //nolint:gosec annotations document the safety invariant
locally (matching the store/extent.go pattern).

multipart.go's `_, _ = NodeToUint32(...)` discards are left in place;
they now suppress a real overflow error, surfacing the latent bug for
the silent-failure-paths follow-up (GH #19 item 5).

GH #19 item 2 / mulga-bm-19.

+++ quic/quicserver/wedge_test.go
@@ -98,7 +98,7 @@ func TestQuicServer_ConcurrentShardWrites_NoWedge(t *testing.T) {
 				Object:     fmt.Sprintf("obj-%d", idx),
 				ObjectHash: hash,
 				ShardSize:  shardSize,
-				ShardIndex: idx,
+				ShardIndex: uint32(idx),
 			}, bytes.NewReader(body))
 			errCh <- perr
 		}(i)
